stashed pop
