+++ routes/apiRoutes.js
@@ -8,30 +8,7 @@ const Helper = require("./utils/helper");
 
 //TODO: use async/await
 // Get all data from current user
-apiRouter.get("/alldata", (req, res, next)=> {
-    const userId = Helper.checkAuth(req.user);
-    if(userId){
-        db.User.findOne({
-            where:{
-                id:userId
-            }
-        }).then(function(userResult) {
-            let stoolPromise = userResult.getStools();
-            let waterPromise = userResult.getWaters();
-
-            Promise.all([stoolPromise,waterPromise]).then(results=>{
-                let result={};
-                result.stool = results[0];
-                result.water = results[1];
-                res.send(result);
-            });
-        }).catch(err=>{
-            next(err);
-        });
-    }else{
-        next(401);
-    }
-});
+apiRouter.get("/alldata", APIController);
 
 //post a new stool log for user
 apiRouter.post("/stool", (req, res, next)=> {
